refactor(web): extract storage form validations to validations.ts
Roughly half of each storage form's fields.ts was validation logic, with
the field vocabulary (types, constants, defaults) making up the rest.
Move the validators and the exported validate function to a dedicated
validations.ts in both partition-form and logical-volume-form.

Each form directory now gives every concern a predictable home:
fields.ts (vocabulary), validations.ts (rules), transformations.ts
(mapping to and from ConfigModel) and queries.ts (data hooks). The
dependency direction stays clean: validations.ts imports from fields.ts,
never the reverse.

To keep form-local code importing its whole vocabulary from one module,
the fields.ts re-export now also covers the shared field-group types.

The fields.test.ts files were renamed to validations.test.ts since the
validate tests were most of their content; the isReusing* tests stay in
new minimal fields.test.ts files.
